+++ project_composer/processors/__init__.py
@@ -1,6 +1,5 @@
 from .base import ComposerProcessor
 from .classes import ClassProcessor
-from .django import DjangoSettingsProcessor, DjangoUrlsProcessor
 from .purge import PurgeProcessor
 from .text import TextContentProcessor
 
@@ -8,8 +7,6 @@ from .text import TextContentProcessor
 __all__ = [
     "ComposerProcessor",
     "ClassProcessor",
-    "DjangoSettingsProcessor",
-    "DjangoUrlsProcessor",
     "PurgeProcessor",
     "TextContentProcessor",
 ]